pyproject.toml
[project]
name = "gudget"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "sqlalchemy>=2.0.41",
]
